HW4:xv6 lazy page allocation

+++ HW3:xv6_system_calls/xv6/trap.c
@@ -13,7 +13,7 @@ struct gatedesc idt[256];
 extern uint vectors[];  // in vectors.S: array of 256 entry pointers
 struct spinlock tickslock;
 uint ticks;
-
+int mappages(pde_t *pgdir, void *va, uint size, uint pa, int perm);
 void
 tvinit(void)
 {
@@ -86,6 +86,21 @@ trap(struct trapframe *tf)
               tf->trapno, cpuid(), tf->eip, rcr2());
       panic("trap");
     }
+	if (tf->trapno == T_PGFLT) {
+		char *mem;
+		uint a;
+
+		a = PGROUNDDOWN(rcr2());
+		mem = kalloc();
+		if (mem == 0) {
+			cprintf("kalloc out of memory!\n");
+			myproc()->killed = 1;	
+			break;
+		}
+		memset(mem, 0, PGSIZE);
+		mappages(myproc()->pgdir, (char*)a, PGSIZE, V2P(mem), PTE_W|PTE_U);
+		break;	
+	}
     // In user space, assume process misbehaved.
     cprintf("pid %d %s: trap %d err %d on cpu %d "
             "eip 0x%x addr 0x%x--kill proc\n",
